[#13] refactor: Refactor BaseEntity to include basic audit fields
