+++ src/main/java/org/dnd/modutimer/common/domain/BaseEntity.java
@@ -1,21 +1,24 @@
-package org.dnd.modutimer.user.application;
+package org.dnd.modutimer.common.domain;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
-import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
 
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
 @EqualsAndHashCode(of = "id", callSuper = false) // Equals()와 Hashcode() 만들어줌
 @Getter
-public class AbstractJpaEntity {
+public class BaseEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     protected Long id;
@@ -23,17 +26,9 @@ public class AbstractJpaEntity {
     @Column(name = "isDeleted", nullable = false, columnDefinition = "BIT default 0")
     protected Boolean isDeleted = false;
 
-    @CreatedBy
-    @Column(name = "created_by", updatable = false, length = 64)
-    protected String createdBy;
-
     @CreatedDate
     @Column(name = "created_at", nullable = false, updatable = false)
     protected LocalDateTime createdAt;
-
-    @LastModifiedBy
-    @Column(name = "last_modified_by", length = 64)
-    protected String lastModifiedBy = "";
 
     @LastModifiedDate
     @Column(name = "last_modified_at", nullable = false)
